Rename project to cellar
Given arsenal already exists [0] and is openstack-related, rename this
project to a new name to avoid confusion.

[0] https://github.com/rackerlabs/arsenal
