--- .gitreview
@@ -1,4 +1,0 @@
-[gerrit]
-host=review.openstack.org
-port=29418
-project=openstack/arsenal.git
