+++ HACKING.rst
@@ -1,4 +1,4 @@
-arsenal Style Commandments
+cellar Style Commandments
 ===============================================
 
 Read the OpenStack Style Commandments http://docs.openstack.org/developer/hacking/
