+++ README.rst
@@ -1,6 +1,6 @@
-=======
-arsenal
-=======
+======
+cellar
+======
 
 Datacenter modelizer.
 
@@ -9,9 +9,9 @@ Please fill here a long description which must be at least 3 lines wrapped on
 Note that this is a hard requirement.
 
 * Free software: Apache license
-* Documentation: http://docs.openstack.org/developer/arsenal
-* Source: http://git.openstack.org/cgit/openstack/arsenal
-* Bugs: http://bugs.launchpad.net/arsenal
+* Documentation: http://docs.openstack.org/developer/cellar
+* Source: http://git.openstack.org/cgit/openstack/cellar
+* Bugs: http://bugs.launchpad.net/cellar
 
 Features
 --------
