+++ cellar/adapters/__init__.py
