+++ cellar/adapters/resource_type_yaml_parser.py
