+++ cellar/core/patch.py
@@ -11,7 +11,7 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations
 # under the License.
-from arsenal import Model
+from cellar import Model
 
 
 class Operation(Model):
