+++ cellar/interfaces/templates/resources.html
@@ -1,6 +1,6 @@
 <html>
 <head>
-    <title>Arsenal overview</title>
+    <title>cellar overview</title>
 </head>
 <body>
 <script src="http://mathieumitchell.com/viz.js">
