+++ cellar/tests/core/test_resource_type_factory.py
@@ -11,7 +11,7 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations
 # under the License.
-from arsenal.core.resource_type import ResourceTypeFactory, ResourceTypeAlreadyExist, UnknownResourceType
+from cellar.core.resource_type import ResourceTypeFactory, ResourceTypeAlreadyExist, UnknownResourceType
 from oslotest import base
