+++ doc/source/index.rst
@@ -1,9 +1,9 @@
-.. arsenal documentation master file, created by
+.. cellar documentation master file, created by
    sphinx-quickstart on Tue Jul  9 22:26:36 2013.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to arsenal's documentation!
+Welcome to cellar's documentation!
 ========================================================
 
 Contents:
