+++ releasenotes/source/index.rst
@@ -1,5 +1,5 @@
 ============================================
- arsenal Release Notes
+ cellar Release Notes
 ============================================
 
 .. toctree::
