+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = arsenal
+name = cellar
 summary = Datacenter modelizer
 description-file =
     README.rst
@@ -21,7 +21,7 @@ classifier =
 
 [files]
 packages =
-    arsenal
+    cellar
 
 [build_sphinx]
 source-dir = doc/source
@@ -32,18 +32,18 @@ all_files = 1
 upload-dir = doc/build/html
 
 [compile_catalog]
-directory = arsenal/locale
-domain = arsenal
+directory = cellar/locale
+domain = cellar
 
 [update_catalog]
-domain = arsenal
-output_dir = arsenal/locale
-input_file = arsenal/locale/arsenal.pot
+domain = cellar
+output_dir = cellar/locale
+input_file = cellar/locale/cellar.pot
 
 [extract_messages]
 keywords = _ gettext ngettext l_ lazy_gettext
 mapping_file = babel.cfg
-output_file = arsenal/locale/arsenal.pot
+output_file = cellar/locale/cellar.pot
 
 [build_releasenotes]
 all_files = 1
